[move] remoteに内容を変更
